+++ src/Enums/PdfLayout.php
@@ -33,7 +33,7 @@ enum PdfLayout: string implements PdfEnumDefaultInterface
     case DEFAULT = '';
 
     /**
-     * Displays the pages in one column.
+     * Displays pages in one column.
      */
     case ONE_COLUMN = 'OneColumn';
 
@@ -43,24 +43,24 @@ enum PdfLayout: string implements PdfEnumDefaultInterface
     case SINGLE_PAGE = 'SinglePage';
 
     /**
-     * Displays the pages in two columns, with odd-numbered pages on the left.
+     * Displays pages in two columns, with odd-numbered pages on the left.
      */
     case TWO_COLUMN_LEFT = 'TwoColumnLeft';
 
     /**
-     * Displays the pages in two columns, with odd-numbered pages on the right.
+     * Displays pages in two columns, with odd-numbered pages on the right.
      */
     case TWO_COLUMN_RIGHT = 'TwoColumnRight';
 
     /**
-     * Displays the pages two at a time, with odd-numbered pages on the left.
+     * Displays two pages at a time, with odd-numbered pages on the left.
      *
      * Require PDF 1.5.
      */
     case TWO_PAGE_LEFT = 'TwoPageLeft';
 
     /**
-     * Displays the pages two at a time, with odd-numbered pages on the right.
+     * Displays two pages at a time, with odd-numbered pages on the right.
      *
      * Require PDF 1.5.
      */